chore: refactor state machine clone logic out as a separate class
